fix everything

+++ aMode.cpp
@@ -1,21 +1,21 @@
 #include "aMode.h"
 
-int add(int num1, int num2)
+double add(double num1, double num2)
 {
     return num1 + num2;
 }
 
-int sub(int num1, int num2)
+double sub(double num1, double num2)
 {
     return num1 - num2;
 }
 
-int divi(int num1, int num2)
+double divi(double num1, double num2)
 {
     return num1/num2;
 }
 
-int mul(int num1, int num2)
+double mul(double num1, double num2)
 {
     return num1*num2;
 }
