+++ aMode.h
@@ -1,9 +1,9 @@
 #ifndef AMODE_H_INCLUDED
 #define AMODE_H_INCLUDED
 
-int add(int num1, int num2);
-int sub(int num1, int num2);
-int divi(int num1, int num2);
-int mul(int num1, int num2);
+double add(double num1, double num2);
+double sub(double num1, double num2);
+double divi(double num1, double num2);
+double mul(double num1, double num2);
 
 #endif // AMODE_H_INCLUDED
